pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "szgmu-schedule-bot"
version = "2.0.0"
description = "SZGMU Schedule Bot - Telegram bot for university schedule management"
readme = "README.md"
license = "MIT"
authors = [
    {name = "Schedule Processor Team"}
]
keywords = ["schedule", "ics", "excel", "telegram-bot", "university"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Education", 
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Education",
    "Topic :: Office/Business :: Scheduling",
]
requires-python = ">=3.10"
dependencies = [
    # Database
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.20.0",
    "greenlet>=3.0.0",  # Required for async SQLAlchemy
    "typing-extensions>=4.9.0",
    "aiogram>=3.0.0",
    "ics>=0.7.2", 
    "openpyxl>=3.1.2",
    "python-dateutil>=2.8.2",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "rich>=13.7.0",
    "loguru>=0.7.0",
    "PyYAML>=6.0.0",
    "click>=8.0.0",
    "greenlet>=3.0.0",  # Required for async SQLAlchemy
    "python-telegram-bot>=20.7",
    "loguru>=0.7.2",
    "anyio>=4.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0", 
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]

[project.scripts]
szgmu-bot = "main:main"

[project.urls]
Homepage = "https://github.com/yourusername/schedule-processor"
Repository = "https://github.com/yourusername/schedule-processor" 
Issues = "https://github.com/yourusername/schedule-processor/issues"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.1",
    "pytest-mock>=3.12.0", 
    "black>=25.1.0",
    "isort>=5.13.2",
    "flake8>=7.1.1",
    "mypy>=1.14.0",
    "pre-commit>=4.0.1",
]


